% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbols.R
\name{getSymbols}
\alias{getSymbols}
\title{Lay gia chung khoan vietNam}
\usage{
getSymbols(symbol, from, to)
}
\arguments{
\item{symbol}{Ma chung khoan, gom 3 ki tu}

\item{from}{ngay bat dau dinh dang yyyy-mm-dd}

\item{to}{ngay ket thuc dinh dang yyyy-mm-dd}
}
\value{
mot data frame chua gia chung khoan va khoi luong
}
\description{
Ham so nay se lay gia chung khoan tu ngay bat dau den ngay ket thuc
}
